Struct pikadick::client_data::ClientData

pub struct ClientData {

    pub shard_manager: Arc<ShardManager>,
    pub nekos_client: NekosClient,
    pub r6stats_client: R6StatsClient,
    pub r6tracker_client: R6TrackerClient,
    pub rule34_client: Rule34Client,
    pub quizizz_client: QuizizzClient,
    pub fml_client: FmlClient,
    pub shift_client: ShiftClient,
    pub reddit_embed_data: RedditEmbedData,
    pub enabled_check_data: EnabledCheckData,
    pub insta_client: Client,
    pub deviantart_client: DeviantartClient,
    pub urban_client: UrbanClient,
    pub xkcd_client: Client,
    pub tic_tac_toe_data: TicTacToeData,
    pub iqdb_client: IqdbClient,
    pub sauce_nao_client: SauceNaoClient,
    pub open_ai_client: Client,
    pub yodaspeak: Client,
    pub tiktok_data: TikTokData,
    pub encoder_task: EncoderTask,
    pub db: Database,
    pub config: Arc<Config>,

}

The ClientData.

Fields
shard_manager: Arc<ShardManager>

The discord shard_manager

nekos_client: NekosClient

The client for nekos

r6stats_client: R6StatsClient

The R6Stats client

r6tracker_client: R6TrackerClient

The r6tracker client

rule34_client: Rule34Client

The rule34 client

quizizz_client: QuizizzClient

The quizizz client

fml_client: FmlClient

The fml client

shift_client: ShiftClient

The shift client

reddit_embed_data: RedditEmbedData

The reddit embed data

enabled_check_data: EnabledCheckData

The enabled check data

insta_client: Client

The insta client data

deviantart_client: DeviantartClient

The deviantart client

urban_client: UrbanClient

The urban dictionary client

xkcd_client: Client

The xkcd client

tic_tac_toe_data: TicTacToeData

The tic tac toe data

iqdb_client: IqdbClient

The iqdb client

sauce_nao_client: SauceNaoClient

The sauce nao client

open_ai_client: Client

The open ai client

yodaspeak: Client

The yodaspeak client

tiktok_data: TikTokData

TikTokData

encoder_task: EncoderTask

Encoder Task

db: Database

The database

config: Arc<Config>

The config

Implementations

impl ClientData

pub async fn init( shard_manager: Arc<ShardManager>, config: Arc<Config>, db: Database ) -> Result<Self>

Init this client data

pub fn generate_cache_stats( &self ) -> BTreeMap<&'static str, BTreeMap<&'static str, f32>>

Generate cache stats Currently, In order for something to show up in cache-stats it must be added here. More automation is desirable in the future.

pub async fn shutdown(&self)

Shutdown anything that needs to be shut down.

Errors are logged to the console, but not returned to the user as it is assumed that they don’t matter in the middle of a shutdown.