pikadick::commands::reddit_embed

Struct RedditEmbedData

pub struct RedditEmbedData {
    reddit_client: Client,
    reddit_tube_client: Client,
    pub cache: TimedCache<(String, String), String>,
    pub video_data_cache: TimedCache<String, Box<GetVideoResponseOk>>,
    random_post_cache: Arc<DashMap<String, Arc<(Instant, Vec<Arc<Link>>)>>>,
}

Fields
reddit_client: Clientreddit_tube_client: Clientcache: TimedCache<(String, String), String>video_data_cache: TimedCache<String, Box<GetVideoResponseOk>>random_post_cache: Arc<DashMap<String, Arc<(Instant, Vec<Arc<Link>>)>>>
Implementations

impl RedditEmbedData

pub fn new() -> Self

Make a new RedditEmbedData.

pub async fn get_original_post( &self, subreddit: &str, post_id: &str, ) -> Result<Box<Link>>

Get the original post from a given subreddit and post id.

This resolves crossposts. Currently only resolves 1 layer.

pub async fn get_video_data(&self, url: &str) -> Result<Box<GetVideoResponseOk>>

Get video data from reddit.tube.

Takes a reddit url.

pub async fn get_video_data_cached( &self, url: &str, ) -> Result<Arc<TimedCacheEntry<Box<GetVideoResponseOk>>>>

Get video data, but using a cache.

pub async fn create_video_url(&self, url: &str) -> Result<Url>

Create a video url for a url to a reddit video post.

pub async fn get_embed_url(&self, url: &Url) -> Result<String>

Get a reddit embed url for a given subreddit and post id

pub async fn try_embed_url( &self, ctx: &Context, msg: &Message, url: &Url, loading_reaction: &mut Option<LoadingReaction>, ) -> Result<()>

Try to embed a url

pub async fn get_random_post(&self, subreddit: &str) -> Result<Option<String>>

Get a random post url for a subreddit

async fn reddit_link_to_embed_url(&self, link: &Link) -> Result<String>

Convert a reddit link to an embed url