Module pikadick::commands::tic_tac_toe

Modules

board

concede

play

tic_tac_toe_renderer

Structs

GamePlayerMention

GameState
A Tic-Tac-Toe game.

InvalidGamePlayer

TicTacToeData
Data pertaining to running tic_tac_toe games

Enums

CreateGameError
Error that may occur while creating a game.

GamePlayer
A player of Tic-Tac-Toe

TryMoveError
Error that may occur while performing a game move.

TryMoveResponse
The response for making a move.

Statics

BOARD_COMMAND

CONCEDE_COMMAND

PLAY_COMMAND

TIC_TAC_TOE_COMMAND

TIC_TAC_TOE_COMMAND_OPTIONS

Functions

tic_tac_toe

Type Definitions

GameStateKey
A GuildId/UserId key to a GameState.

ShareGameState
A GameState that is wrapped in a mutex and sharable via a rc’ed ptr.