pikadick::database

Struct Database

pub struct Database {
    db: Database,
}

The database

Fields
db: Database
Implementations

impl Database

pub async fn set_disabled_command( &self, id: GuildId, cmd: &str, disable: bool, ) -> Result<bool>

Disables or enables a command.

Returns

Returns the old setting

pub async fn is_command_disabled(&self, id: GuildId, name: &str) -> Result<bool>

Check if a command is disabled

impl Database

pub async fn store_get<P, K, V>(&self, prefix: P, key: K) -> Result<Option<V>>
where P: AsRef<[u8]>, K: AsRef<[u8]>, V: DeserializeOwned,

Get a key from the store

pub async fn store_put<P, K, V>( &self, prefix: P, key: K, value: V, ) -> Result<()>
where P: AsRef<[u8]>, K: AsRef<[u8]>, V: Serialize,

Put a key in the store

pub async fn store_update<P, K, V, U>( &self, prefix: P, key: K, update_func: U, ) -> Result<()>
where P: AsRef<[u8]>, K: AsRef<[u8]>, V: Serialize + DeserializeOwned, U: FnOnce(Option<V>) -> V + Send + 'static,

Get and Put a key in the store in one action, ensuring the key is not changed between the commands.

impl Database

pub async fn set_reddit_embed_enabled( &self, guild_id: GuildId, enabled: bool, ) -> Result<bool>

Enable or disable reddit embeds.

Returns

Returns the old value

pub async fn get_reddit_embed_enabled(&self, guild_id: GuildId) -> Result<bool>

Get the reddit embed setting.

impl Database

pub async fn create_tic_tac_toe_game( &self, guild_id: MaybeGuildString, author: TicTacToePlayer, author_team: Team, opponent: TicTacToePlayer, ) -> Result<TicTacToeGame, TicTacToeCreateGameError>

Create a new tic-tac-toe game

pub async fn try_tic_tac_toe_move( &self, guild_id: MaybeGuildString, player: TicTacToePlayer, move_index: u8, ) -> Result<TicTacToeTryMoveResponse, TicTacToeTryMoveError>

Try to make a tic-tac-toe move

pub async fn get_tic_tac_toe_game( &self, guild_id: MaybeGuildString, player: TicTacToePlayer, ) -> Result<Option<TicTacToeGame>>

Try to get a tic-tac-toe game by guild and player

pub async fn concede_tic_tac_toe_game( &self, guild_id: MaybeGuildString, player: UserId, ) -> Result<Option<TicTacToeGame>>

Try to concede a Tic-Tac-Toe game.

Returns

Returns the game if it existed

pub async fn get_tic_tac_toe_score( &self, guild_id: MaybeGuildString, player: UserId, ) -> Result<TicTacToeScore>

Get the userโs Tic-Tac-Toe scores

pub async fn get_top_tic_tac_toe_scores( &self, guild_id: MaybeGuildString, ) -> Result<Vec<TicTacToeTopPlayerScore>>

Get the top Tic-Tac-Toe scores for the current server

impl Database

pub async fn set_tiktok_embed_flags( &self, guild_id: GuildId, set_flags: TikTokEmbedFlags, unset_flags: TikTokEmbedFlags, ) -> Result<(TikTokEmbedFlags, TikTokEmbedFlags)>

Set the flags for tiktok embeds.

Returns

Returns the old flags and new flags in a tuple in that order.

pub async fn get_tiktok_embed_flags( &self, guild_id: GuildId, ) -> Result<TikTokEmbedFlags>

Get the tiktok embed flags.

impl Database

pub async unsafe fn new<P>(path: P, create_if_missing: bool) -> Result<Self>
where P: Into<Utf8PathBuf>,

Safety

This must be called before any other sqlite functions are called.

pub unsafe fn blocking_new<P>(path: P, create_if_missing: bool) -> Result<Self>
where P: AsRef<Utf8Path>,

Make a new Database in a blocking manner.

Safety

This must be called before any other sqlite functions are called.

async fn access_db<F, R>(&self, func: F) -> Result<R>
where F: FnOnce(&mut Connection) -> R + Send + 'static, R: Send + 'static,

Access the db

pub async fn close(&self) -> Result<()>

Close the db