Module pikadick::database::tic_tac_toe

Enums

TicTacToeCreateGameError

Error that may occur while creating a tic-tac-toe game

TicTacToeTryMoveError

Error that may occur while performing a tic-tac-toe move

TicTacToeTryMoveResponse

The response for making a tic-tac-toe move

Constants

CHECK_IN_TIC_TAC_TOE_GAME_SQL 🔒

CREATE_DEFAULT_SCORE_TIC_TAC_TOE_SQL 🔒

CREATE_TIC_TAC_TOE_GAME_SQL 🔒

DELETE_TIC_TAC_TOE_GAME_SQL 🔒

GET_TIC_TAC_TOE_GAME_SQL 🔒

GET_TIC_TAC_TOE_SCORE_SQL 🔒

GET_TOP_TIC_TAC_TOE_SCORES_SQL 🔒

INCREMENT_CONCEDES_SCORE_TIC_TAC_TOE_SQL 🔒

INCREMENT_LOSSES_SCORE_TIC_TAC_TOE_SQL 🔒

INCREMENT_TIES_SCORE_TIC_TAC_TOE_SQL 🔒

INCREMENT_WINS_SCORE_TIC_TAC_TOE_SQL 🔒

UPDATE_TIC_TAC_TOE_GAME_SQL 🔒

Functions

create_user_score_data 🔒

Try to make a user’s score data

delete_tic_tac_toe_game 🔒

get_tic_tac_toe_game 🔒

set_draw_tic_tac_toe_game 🔒

Set a tic-tac-toe game as a draw as part of a larger transaction, consuming it.

set_win_tic_tac_toe_game 🔒

Set a tic-tac-toe game as a win as part of a larger transaction, consuming it.

update_tic_tac_toe_game 🔒