Module pikadick::util

Modules

ascii_table 🔒

encoder_task 🔒

loading_reaction 🔒

timed_cache 🔒

Structs

AsciiTable

An ascii table

EncoderTask

A task to re-encode things

LoadingReaction

This type attaches to a message and displays a loading sign until send_ok or send_err are called, where it then displays a check or an X respectively. If neither are called, send_err is called automatically from the destructor. All functions are not async and can only be used from a tokio runtime context. Errors are silently ignored.

TimedCache

A cache with entries that “expire” after a per-cache time limit

TimedCacheEntry

Statics

URL_REGEX 🔒

Source: https://urlregex.com/

Functions

extract_urls

Get an iterator over urls in text.