pikadick::util

Module loading_reaction

Structs

LoadingReaction

This type attaches to a message and displays a loading sign until send_ok or send_err are called, where it then displays a check or an X respectively. If neither are called, send_err is called automatically from the destructor. All functions are not async and can only be used from a tokio runtime context. Errors are silently ignored.

Constants

ERR_EMOJI 🔒

LOADING_EMOJI 🔒

OK_EMOJI 🔒