Module pikadick::util

Structs

LoadingReaction
This type attaches to a message and displays a loading sign until send_ok or send_err are called, where it then displays a check or an X respectively. If neither are called, send_err is called automatically from the destructor. All functions are not async and can only be used from a tokio runtime context. Errors are silently ignored.

TimedCache
A cache with entries that “expire” after a per-cache time limit

TimedCacheEntry

TimedCacheInner

Constants

DEFAULT_EXPIRE_TIME
10 minutes

ERR_EMOJI

LOADING_EMOJI

OK_EMOJI